requirements.txt
recipe-scrapers>=15.0.0
requests>=2.31.0
rich>=13.0.0
yt-dlp>=2023.10.0
openai-whisper>=20231117
